tdf#114704 use of float as intermediate causes out by one on large offsets
table offset integer value of 21281769 is correctly read, but on cast to float
it is represented as 21281768 and we're off by one when cast back to integer
later

+++ vcl/source/fontsubset/cff.cxx
@@ -31,7 +31,7 @@ typedef long long S64;
 
 typedef sal_Int32 GlyphWidth;
 
-typedef float RealType;
+typedef double RealType;
 typedef RealType ValType;
 #include <vector>
